Support a non-default field length
This plugin uses the plugin_instance field to encode dimension data for
RabbitMQ metrics. The default maximum length of plugin_instance in
collectd is 63 characters. In some cases, the dimension data may
exceed 63 characters, causing data to be truncated.

It is possible that some users will want to recompile collectd to use a
larger field length (DATA_MAX_NAME_LEN in plugin.h). This commit adds a
new configuration option, FieldLength, which enables sending a greater
amount of RabbitMQ dimension data to SignalFx.

This option should _not_ be used unless you are running with a
custom-compiled build of collectd that has a modified DATA_MAX_NAME_LEN.

An additional change made in this commit is the "name" dimension is
given protection priority when truncating dimension data beyond the
field length. This may cause some RabbitMQ-specific dimension names to
change when switching to this version from a prior commit.