Disallow negative counters
The collectd python plugin casts counter values to unsigned long in C.
This triggers an OverflowError or TypeError depending on the
python version when posting a counter with a negative value. This commit
skips reporting counter metrics with negative values.

Also increased the default http timeout for RabbitMQ API calls for
systmes under heavy load or with a large number of connections and
queues.