Add verbosity levels for metrics
This commit adds an additional level of control for which metrics
generated by RabbitMQ are sent to SignalFx. It introduces a new config
file setting, VerbosityLevel. VerbosityLevel is an integer that
represents which metrics are reported:

1: Only the most commonly-used metrics are reported
2: Additonal metrics useful for debugging are reported
3: All available metrics are reported

The default value is verbosity level 1. The setting can be modified in
the 10_rabbitmq.conf file in the collectd managed_config directory.